Moved animations to own file.
